requirements.txt
Django==4.2.11
djangorestframework==3.15.1
nltk==3.8.1
requests==2.31.0
celery==5.3.5
redis==5.0.1
Markdown==3.6
eventlet==0.35.1
gunicorn==21.2.0
psycopg2-binary==2.9.9